moved Polyblade impl to main
