+++ src/scene/mod.rs
@@ -2,18 +2,15 @@ mod camera;
 mod pipeline;
 mod polygon;
 use crate::polyhedra::PolyGraph;
-use crate::{wgpu, Polyblade, RGB};
+use crate::{wgpu, RGB};
 use camera::Camera;
-use iced::mouse;
 use iced::time::Duration;
-use iced::widget::shader;
-use iced::Rectangle;
 pub use pipeline::*;
 use std::f32::consts::PI;
 use std::time::Instant;
 use ultraviolet::Mat4;
 
-use polygon::Polygon;
+pub use polygon::Polygon;
 
 pub struct AppState {
     pub polyhedron: PolyGraph,
@@ -70,44 +67,5 @@ impl AppState {
                 * Mat4::from_rotation_x(time / PI)
                 * Mat4::from_rotation_y(time / PI * 1.1);
         }
-    }
-}
-
-impl<Message> shader::Program<Message> for Polyblade {
-    type State = ();
-    type Primitive = Polygon;
-
-    /* fn update(
-        &self,
-        _state: &mut Self::State,
-        event: shader::Event,
-        _bounds: Rectangle,
-        _cursor: mouse::Cursor,
-        shell: &mut Shell<'_, Message>,
-    ) -> (event::Status, Option<Message>) {
-        match event {
-            /* shader::Event::Mouse(_) => {}
-            shader::Event::Touch(_) => {}
-            shader::Event::Keyboard(_) => {} */
-            shader::Event::RedrawRequested(time) => {
-                println!("redraw requested11");
-                (event::Status::Captured, None)
-            }
-            _ => (event::Status::Ignored, None),
-        }
-    } */
-
-    fn draw(
-        &self,
-        _state: &Self::State,
-        _cursor: mouse::Cursor,
-        _bounds: Rectangle,
-    ) -> Self::Primitive {
-        Polygon::new(
-            &self.state.polyhedron,
-            &self.state.palette,
-            &self.state.transform,
-            &self.state.camera,
-        )
     }
 }
